fix: move inline tests to mod_tests.rs and fix stale comment
Address CLAUDE.md convention violations from PR #1624 review:
- Move inline #[cfg(test)] block from mod.rs to separate mod_tests.rs
  (matches the existing pattern of 6 sibling *_tests.rs files)
- Update stale comment in draw_channel_switcher_overlay to mention
  @peer format alongside #channel-name
